requirements/prod.txt
Django==2.1.7
djangorestframework==3.11.0
djongo==1.2.38
isort==4.3.21
pylint==2.4.4
pymongo==3.10.0
PyJWT==1.4.0
